Fix information computation with weighted samples.
* Change the type of `environment.n_eff` and `InfoBlock.n_samples` from
int to double.

* Use double values for the computation of entropy (n * log(n)), use
rounded values only for the computation of complexity terms.

+++ src/environment.h
@@ -26,7 +26,7 @@ struct Environment {
   vector<int> is_continuous;
   vector<int> levels;
   vector<int> has_na;
-  int n_eff;
+  double n_eff;
   vector<Node> nodes;
   Grid2d<Edge> edges;
   bool orientation = false;
